requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
